Use custom keyboards in surveys
Refactored survey schema to add store option to indicate where in the context should be stored the answer, no matter whether it is an intent, an entity or a message.

Support transitions on exact message, without relying on wit.ai if there is a match.

Support options rendered as custom keyboards.

Fixes #41.