Improved texts for price queries
If the user asks for a price that has a hit on his marketplace, then return it.

Otherwise, if there are hits on his region, then return them stating "Prices in REGION are X in MKT1, Y in MKT2, etc".

Otherwise, return prices in the country stating "Prices in COUNTRY are X in MKT1 (ADM1), Y in MKT2 (ADM2), etc".

Fixes #66.