#57 Do not ask for gps again if we have recent gps data
If a user provided recent GPS but it couldn't be matched to a known mkt, we were starting the geolocation process from scratch, which seemed unnatural (why would the second attempt succedd?).
We are now telling the user that a known marketplace couldn't be found and giving him information on how to reset location.